requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==2.2.18
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-extensions==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
ipython==7.20.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.8.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
